Split up impl.go into schema, lineage, instance
